+++ Domain/Projects/Project_Stopped.cs
@@ -38,7 +38,7 @@ namespace Domain.Projects
         public long ProposedPrice { get; set; }
         //مشارکت
         //میزان مشارکت مورد نیاز
-        public int AmountOfParticipationRequired { get; set; }
+        public long AmountOfParticipationRequired { get; set; }
         //مشارکت
         //مدت زمان غیرفعال بودن طرح (سال)
         public int DurationInactivityYear { get; set; }
