+++ WebApi/ModelsAndDtoes/Project_Stopped/CreateProject_StoppedApiDto.cs
@@ -50,7 +50,7 @@ namespace WebApi.ModelsAndDtoes.Project_Stopped
         public long ProposedPrice { get; set; }
         //مشارکت
         //میزان مشارکت مورد نیاز
-        public int AmountOfParticipationRequired { get; set; }
+        public long AmountOfParticipationRequired { get; set; }
         //مشارکت
         //مدت زمان غیرفعال بودن طرح (سال)
         public int DurationInactivityYear { get; set; }
